# openapi.yaml
openapi: 3.1.0
info:
  title: Health Assistant
  version: 0.3.0
  description: >-
    LAN-only conversational text-to-SQL service over the Semantic Layer (gold views +
    sleep_score/device/profile). The model writes read-only SQL, the execute contract
    bounds it (10 s timeout, 500-row cap, EXPLAIN dry-run, ≤2 repairs), and every
    exchange is logged to pipeline.chat_log. Served by FastAPI; the live contract is
    also generated by the app at /openapi.json.
servers:
  - url: http://localhost:8000
    description: local development (LAN deployment mirrors the shape exactly)
tags:
  - name: chat
    description: streaming conversation end to end
  - name: feedback
    description: one thumbs rating per logged exchange
  - name: meta
    description: the app page and the API contract itself
paths:
  /:
    get:
      tags: [meta]
      summary: Single-page chat UI (no build step, no framework)
      responses:
        "200":
          description: the frontend/index.html page, served with its local css/js assets
          content:
            text/html:
              schema:
                type: string
  /openapi.yaml:
    get:
      tags: [meta]
      summary: This contract, served as YAML by the app itself
      responses:
        "200":
          description: the OpenAPI document
          content:
            application/yaml:
              schema:
                type: string
  /api/chat:
    post:
      tags: [chat]
      summary: Ask a question; the answer streams back as Server-Sent Events
      operationId: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                  minLength: 1
                  description: the health-data question in plain language
                session_id:
                  type: string
                  default: anon
                  description: >-
                    stable per-conversation id (in-memory context for follow-ups;
                    also sent to the provider as x-opencode-session)
      responses:
        "200":
          description: >-
            text/event-stream with these events:
            session (id), sql (executed query + row count + truncation/repair flags),
            token (streamed answer text), outcome (answered|refused|error + Chat Log id)
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                session:
                  value: "event: session\ndata: {\"session_id\":\"s-1\"}\n\n"
                outcome:
                  value: "event: outcome\ndata: {\"outcome\":\"answered\",\"text\":\"496\",\"row_count\":1,\"log_id\":3}\n\n"
        "422":
          description: empty question (validation)
        "503":
          description: assistant offline — no usable LLM provider configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "assistant offline: LLM_BASE_URL is not set — point it at the OpenAI-compatible gateway (Ollama http://localhost:11434/v1, the BMF endpoint, or the OpenCode gateway). LLM_PROVIDER=stub runs the deterministic demo provider instead."
  /api/feedback:
    post:
      tags: [feedback]
      summary: Rate one exchange (👍/👎) — exactly once per Chat Log row
      operationId: feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [log_id, thumbs]
              properties:
                log_id:
                  type: integer
                  description: the Chat Log row id from the outcome event
                thumbs:
                  type: string
                  enum: [up, down]
      responses:
        "200":
          description: rating recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    const: true
        "404":
          description: row not found or already rated
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        "422":
          description: invalid thumbs value or missing fields